feat(auth): Allows a user to Signup, SignIn and SignOut
This commit enables a user to signup for the service using
their Andela Email. It leverages devise, omniauth-google-oauth2
gems.

It adds the welcome controller which is responsible for signin,
and the home controller which is the landing page after a user
signed in.

A welcome template has been added to provide for the custom
landing page.

[Delivers #159559077, ]